requirements.txt
flask
gunicorn
transformers
torch
mysql-connector-python
tiktoken
protobuf
